First step of WS secure cookies, refactored logic into 1 spot & invoke call in Websocket_Server
